#CKFK20221108 Resoluciones

+++ app/src/main/java/com/dts/tom/Transacciones/CambioUbicacion/frm_cambio_ubicacion_ciega.java
@@ -2493,10 +2493,10 @@ public class frm_cambio_ubicacion_ciega extends PBase {
 
                 gl.IdResolucionLpOperador = resolucionActivaLpByBodega.IdResolucionlp;
 
-                float pLpSiguiente = resolucionActivaLpByBodega.Correlativo_Actual +1;
-                float largoMaximo = String.valueOf(resolucionActivaLpByBodega.Correlativo_Final).length();
+                long pLpSiguiente = resolucionActivaLpByBodega.Correlativo_Actual +1;
+                int largoMaximo = String.valueOf(resolucionActivaLpByBodega.Correlativo_Final).length();
 
-                int intLPSig = (int) pLpSiguiente;
+                long intLPSig = (long) pLpSiguiente;
                 int MaxL = (int) largoMaximo;
 
                 //#CKFK20220410 Reemplacé el código de arriba por esta línea
